// MotarMarketProAngular/src/app/models/sold-vehicle.model.ts
export interface SoldVehicle {
  sold_id?: number;
  vehicle_id: number;
  vehicle_type: string;
  manufacturer: string;
  model: string;
  fuel_type: string;
  ownership: string;
  reg_data: string;
  veh_number: string;
  price: number;
  sold_date: string;
  buyer_name: string;
  buyer_contact: string;
}